requirements.txt
Django>=5.1,<6
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
psycopg2-binary>=2.9
celery[redis]>=5.3
gunicorn>=22
twilio>=9
anthropic>=0.40
whitenoise>=6.7
dj-database-url>=2.1
